pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "wagapi"
version = "0.1.0"
description = "CLI client for the Wagtail Write API, optimised for LLM orchestration"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "click>=8.0",
    "httpx>=0.24",
    "markdown-it-py>=3.0",
    "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "respx>=0.20",
    "pytest-httpx>=0.20",
]

[project.scripts]
wagapi = "wagapi.cli:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
